Integrate LongRequests and Productionalize SessionViewer/Summary
This commit became much larger than a commit normally should. It includes
the integration of the sp_XR_LongRequests procedure, the creation of a
UserCollectionOptions table and the associated modifications to
user-initiated runs of AutoWho.Collector, changes to the PostProcessor
(previously just a "resolve waits" proc), and many fixes for
SessionSummary and SessionViewer. The help documentation has also been
significantly reworked.