Rework AutoWho framework to capture/display current DMV data
This commit completes implementation of Issue #11, the modification of
the AutoWho framework to support the capture and display of current DMV
data rather than just historical DMV data. All AutoWho tables have been
modified to indicate which area of ChiRho initiated the AutoWho
collection, and procs (both internal and user-facing) have been modified
to support the changes. As of this commit, the ability to capture current
DMV data either in a one-off capture mode or a custom polling trace mode
has been fully implemented.