Add @units parameter/functionality, break out resource metrics
The metrics for LongRequests need to be broken out into more informative
and logical divisions, and the user needs to be able to see the data in
different units. This commit implements a new @units parameter and breaks
out things like "writes" into max/min/first/last/deltas/avg columns.